refactor(src): adjust naming imports - I35
